moved aetherfeatures to registry and added flower and berry generation
