+++ src/main/resources/data/aether/worldgen/biome/aether_skylands.json
@@ -44,6 +44,7 @@
     [
       "minecraft:patch_tall_grass_2",
       "minecraft:patch_grass_plain",
+      "aether:aether_skylands_flowers",
       "aether:tree_skyroot",
       "aether:tree_golden_oak",
       "aether:spring_water",
